chatbot/requirements.txt

langchain-core
langchain-community
streamlit
requests
openai